Modify readFile method
Method calls addToTable method of HashTable and passes the sorted word
as the key and the word as the value. By doing this, all anagrams will
have the same key since every anagram has the same sorted string. Every
anagram will therefore be stored in the same bucket of the hash table
so that they can quickly be retrieved.